requirements.txt
# AI Study Companion - Python RAG Backend Dependencies
# Install with: pip install -r requirements.txt

# -- Core LangChain & Agents -------------------------------------------------
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-google-genai>=0.2.0
langchain-openai>=0.2.0

# -- Vector Database ----------------------------------------------------------
faiss-cpu>=1.7.4

# -- PDF & Image Parsing (OCR) ------------------------------------------------
pypdf>=4.0.0
pytesseract>=0.3.10
Pillow>=10.0.0

# -- Repository Analysis ------------------------------------------------------
GitPython>=3.1.30

# -- Embeddings & ML (Heavy - Commented for Render Free Tier) -----------------
# sentence-transformers>=3.0.0
# torch>=2.0.0
google-generativeai>=0.7.0

# -- API Server ---------------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# -- Utilities ----------------------------------------------------------------
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
esprima
firebase-admin>=6.5.0
